renamed test file, used relative path
